Correctly deal with None values
Currently "mm2li import-orimatis" sets "tdate" to the "None" string in the transfo POST request when there is no calibration date in the orimatis XML document. "None" is not a valid date, so the HTTP API returns an error.

Similarly, "mm2li import-orimatis" sets "name" to the "None" string in the sensor POST request when there is no sensor name in the orimatis XML document. "None" is a valid string, so there is no error in this case. But using "None" as the default sensor name doesn't make sense.

This commit fixes both problems, by correctly dealing with None values.

Fixes #16