Add --filename-pattern argument to import-autocal
Currently, import-autocal, when run with no specific argument, will create a camera sensor with the autocal filename as the sensor name. For example with

    li3ds import-autocal data/AutoCal_Foc-12000_Cam-Caml024_20161205a.xml

a camera sensor named "AutoCal_Foc-12000_Cam-Caml024_20161205a.xml" will be created.

To provide a specific sensor name the "--sensor" argument can be used. For example:

    li3ds import-autocal -s 024 data/AutoCal_Foc-12000_Cam-Caml024_20161205a.xml

But that means one "import-autocal" command per AutoCal file to import, which is not very convenient when many AutoCal files need to be imported.

This commit adds a --filename-pattern argument. When a file name pattern is provided 1) files whose names do not match the pattern are skipped, and 2) the sensor name can be read from the AutoCal file name. For example:

    li3ds import-autocal -p 'AutoCal_Foc-\d+_Cam-Caml(?P<sensor_name>\d+)_20161205a.xml data/AutoCal_*.xml