Make import-extcalib attach camera referentials to camera sensors
Before this commit camera referentials created by import-extcalib were attached to the sensor group. With this commit they are attached to a camera sensor. In this way import-extcalib will not create new camera referentials in the li3ds database if they already exist (for example because they were previously created by import-autocal).